Register new devices on the Knot cloud integration
Whenever a new device is created, send a registration request to the knot cloud

If knot cloud returns that the device has already been registered, send an unregistration request to knot cloud

Whenever the knot cloud returns an unregistration confirmation message, check if the device has a token, if it does, it must be deleted. Otherwise,the absence of a token shows that the device was registered in the knot cloud, but did not have its token saved in the chirpstack, so it needs to resubmit a registration request to get a valid token.